+++ packages/experimental/CHANGELOG.md
@@ -1,1 +1,22 @@
 # @llamaindex/experimental
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [41210df]
+- Updated dependencies [137cf67]
+- Updated dependencies [259c842]
+  - llamaindex@0.2.1
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [d2e8d0c]
+- Updated dependencies [aefc326]
+- Updated dependencies [484a710]
+- Updated dependencies [d766bd0]
+- Updated dependencies [dd95927]
+- Updated dependencies [bf583a7]
+  - llamaindex@0.2.0